# core/contracts/openapi/v1.yaml
openapi: 3.1.0
info:
  title: LLPlayerNext Local API
  version: 1.0.0
  description: Versioned loopback-only application-service adapter. Playback control is excluded.
servers:
  - url: http://127.0.0.1:{port}
    variables:
      port:
        default: "0"
security:
  - bearerAuth: []
paths:
...
  /v1/media:
    post:
      operationId: registerMedia
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterMedia"
      responses:
        "200":
          description: Idempotently registered media
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaItem"
        default:
          $ref: "#/components/responses/Error"
  /v1/lltimeline/import:
    post:
      operationId: importLLTimeline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LLTimelineDocument"
      responses:
        "200":
          description: Imported LLTimeline JSON v1 resource document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubtitleTrack"
  /v1/media/{media_id}:
    parameters:
      - $ref: "#/components/parameters/MediaId"
    get:
      operationId: readMedia
      responses:
        "200":
          description: Media metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaItem"
        default:
          $ref: "#/components/responses/Error"
  /v1/media/{media_id}/progress:
    parameters:
      - $ref: "#/components/parameters/MediaId"
    get:
      operationId: readProgress
      responses:
        "200":
          description: Last saved position
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Progress"
        default:
          $ref: "#/components/responses/Error"
    put:
      operationId: updateProgress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Progress"
      responses:
        "200":
          description: Saved position
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Progress"
        default:
          $ref: "#/components/responses/Error"
  /v1/media/{media_id}/subtitles:
    parameters:
      - $ref: "#/components/parameters/MediaId"
    get:
      operationId: mediaSubtitles
      responses:
        "200":
          description: Subtitle resources associated with a media item
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SubtitleTrack" }
    post:
      operationId: importSubtitle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportSubtitle"
      responses:
        "200":
          description: Complete normalized subtitle timeline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubtitleTrack"
        default:
          $ref: "#/components/responses/Error"
  /v1/media/{media_id}/lltimeline/import:
    parameters:
      - $ref: "#/components/parameters/MediaId"
      - name: allow_mismatch
        in: query
        required: false
        schema: { type: boolean }
        description: Allow user-confirmed attachment when the resource media fingerprint differs.
    post:
      operationId: importLLTimelineForMedia
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LLTimelineDocument"
      responses:
        "200":
          description: Imported and attached LLTimeline JSON v1 resource document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubtitleTrack"
        default:
          $ref: "#/components/responses/Error"
  /v1/subtitles/{track_id}:
    parameters:
      - name: track_id
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: readSubtitle
      responses:
        "200":
          description: Complete normalized subtitle timeline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubtitleTrack"
        default:
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteSubtitle
      responses:
        "200":
          description: Deleted subtitle resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubtitleTrack"
        default:
          $ref: "#/components/responses/Error"
  /v1/subtitles/{track_id}/archive:
    parameters:
      - name: track_id
        in: path
        required: true
        schema: { type: string }
    post:
      operationId: archiveSubtitle
      responses:
        "200":
          description: Archived subtitle resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubtitleTrack"
        default:
          $ref: "#/components/responses/Error"
  /v1/subtitles/{track_id}/restore:
    parameters:
      - name: track_id
        in: path
        required: true
        schema: { type: string }
    post:
      operationId: restoreSubtitle
      responses:
        "200":
          description: Restored subtitle resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubtitleTrack"
        default:
          $ref: "#/components/responses/Error"
  /v1/subtitles/{track_id}/language:
    parameters:
      - name: track_id
        in: path
        required: true
        schema: { type: string }
    patch:
      operationId: updateTrackLanguage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTrackLanguage"
      responses:
        "200":
          description: Updated subtitle learning language
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubtitleTrack"
        default:
          $ref: "#/components/responses/Error"
  /v1/subtitles/{track_id}/export:
    get:
      operationId: exportSubtitle
      parameters:
        - { name: track_id, in: path, required: true, schema: { type: string } }
        - { name: format, in: query, schema: { type: string, default: srt } }
      responses:
        "200": { description: Generated SRT subtitle }
...
  /v1/pronunciation/lookup:
    get:
      operationId: pronunciationLookup
      parameters:
        - { name: word, in: query, required: true, schema: { type: string } }
      responses:
        "200": { description: Canonical word pronunciation and variants }
...
  /v1/subtitles/{track_id}/word-timelines:
    get:
      operationId: trackWordTimelines
      responses:
        "200":
          description: Versioned word timeline resources for a subtitle track
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/WordTimeline" }
    post:
      operationId: createTrackWordTimeline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWordTimeline"
      responses:
        "200":
          description: Created word timeline resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WordTimeline"
...
  /v1/subtitles/{track_id}/chunk-timelines:
    get:
      operationId: trackChunkTimelines
      responses:
        "200":
          description: Versioned chunk timeline resources for a subtitle track
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ChunkTimeline" }
    post:
      operationId: generateChunkTimeline
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateChunkTimeline"
      responses:
        "200":
          description: Generated chunk timeline candidate from the active word timeline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChunkTimeline"
...
  /v1/lexical-entries/batch:
    post:
      operationId: readLexicalEntries
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BatchLexicalEntries" }
      responses:
        "200":
          description: Existing lexical entries for normalized forms
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/LexicalEntry" } }
  /v1/lexical-entries:
    get:
      operationId: listLexicalEntries
      parameters:
        - { name: language, in: query, schema: { type: string, default: en } }
        - { name: kind, in: query, schema: { enum: [word, phrase] } }
        - { name: status, in: query, schema: { $ref: "#/components/schemas/LearningStatus" } }
        - { name: search, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0 } }
      responses:
        "200":
          description: Unified word and phrase learning assets
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/LexicalEntryDetails" } }
    put:
      operationId: upsertLexicalEntry
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpsertLexicalEntry" }
      responses:
        "200":
          description: Updated lexical learning asset
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LexicalEntryDetails" }
  /v1/lexical-entries/{id}:
    get:
      operationId: lexicalEntryDetails
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Lexical asset details
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LexicalEntryDetails" }
  /v1/lexical-entries/{id}/learning-content:
    put:
      operationId: updateLexicalLearningContent
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateLexicalLearningContent" }
      responses:
        "200":
          description: Updated lexical learning content
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LexicalEntryDetails" }
  /v1/lexical-observations:
    post:
      operationId: createLexicalObservation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateLexicalObservation" }
      responses:
        "200":
          description: Context-specific listening observation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LexicalObservation" }
        "204":
          description: Observation cleared
  /v1/lexical-normalization:
    post:
      operationId: normalizeLexical
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NormalizeLexicalRequest" }
      responses:
        "200":
          description: Versioned lexical normalization result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LexicalNormalization" }
  /v1/lexical-normalization/correct:
    post:
      operationId: correctLemma
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CorrectLemmaRequest" }
      responses:
        "200":
          description: Persisted user lemma correction
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LexicalNormalization" }
        "409": { $ref: "#/components/responses/Error" }
  /v1/sentences/{sentence_id}/phrase-candidates:
    get:
      operationId: phraseCandidates
      parameters:
        - { name: sentence_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Phrase candidates requiring user confirmation
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/PhraseCandidate" } }
...
  /v1/learning-resources/{id}/install:
    post:
      operationId: installLearningResource
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Installed and verified learning resource
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LearningResource" }
  /v1/learning-resources/{id}:
    delete:
      operationId: removeLearningResource
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Removed learning resource
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LearningResource" }
  /v1/subtitle-search:
    post:
      operationId: searchSubtitles
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubtitleSearchRequest" }
      responses:
        "200":
          description: Subtitle provider search results
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/SubtitleSearchResult" } }
        "401": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /v1/subtitle-search/download:
    post:
      operationId: downloadSubtitle
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubtitleDownloadRequest" }
      responses:
        "200": { description: Downloaded subtitle bytes }
...
  /v1/vocabulary/import:
    post:
      operationId: importVocabulary
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VocabularyAssetBundle" }
      responses:
        "200": { description: Assets merged }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    MediaId:
      name: media_id
      in: path
      required: true
      schema: { type: string }
  responses:
    Error:
      description: Structured error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    RegisterMedia:
      type: object
      required: [path, fingerprint, title, kind]
      properties:
        path: { type: string }
        fingerprint: { type: string }
        title: { type: string }
        kind: { enum: [video, audio] }
        duration_ms: { type: [integer, "null"], minimum: 0 }
    MediaItem:
      type: object
      required: [id, path, fingerprint, title, kind, duration, availability, created_at_ms, updated_at_ms]
      properties:
        id: { type: string }
        path: { type: string }
        fingerprint: { type: string }
        title: { type: string }
        kind: { enum: [video, audio] }
        duration: { type: [integer, "null"], minimum: 0 }
        availability: { enum: [available, missing, archived] }
        created_at_ms: { type: integer, minimum: 0 }
        updated_at_ms: { type: integer, minimum: 0 }
    Progress:
      type: object
      required: [position_ms]
      properties:
        position_ms: { type: [integer, "null"], minimum: 0 }
...